requirements.txt
Django==2.0.3
django-webpack-loader==0.6.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
mysqlclient==1.3.12
PyJWT==1.6.0
pytz==2018.3
